LPD-3209 Update file path for commerce test modules
