+++ modules/test/playwright/pages/commerce/commerce-product-content-search-web/specificationFacetsPage.ts
@@ -5,7 +5,7 @@
 
 import {Locator, Page} from '@playwright/test';
 
-import {CommerceLayoutsPage} from './commerceLayoutsPage';
+import {CommerceLayoutsPage} from '../commerceLayoutsPage';
 
 export class SpecificationFacetsPage {
 	readonly addSearchOptionsLabel: Locator;
@@ -13,12 +13,14 @@ export class SpecificationFacetsPage {
 	readonly addWidgetButton: Locator;
 	readonly layoutsPage: CommerceLayoutsPage;
 	readonly page: Page;
+	readonly pageLabel: Locator;
 	readonly pageTitle: Locator;
 	readonly panelList: Locator;
 	readonly searchFormInput: Locator;
 	readonly searchOptionsAllowEmptySearchesInput: Locator;
 	readonly searchOptionsConfigurationEditButton: Locator;
 	readonly searchOptionsConfigurationSaveButton: Locator;
+	readonly selectSpecificationFacetPageInput: Locator;
 
 	constructor(page: Page) {
 		this.addSearchOptionsLabel = page
@@ -32,6 +34,9 @@ export class SpecificationFacetsPage {
 		this.addWidgetButton = page.getByTestId('add');
 		this.layoutsPage = new CommerceLayoutsPage(page);
 		this.page = page;
+		this.pageLabel = page
+			.getByTestId('layoutHref')
+			.getByLabel('Specification Facet Page');
 		this.pageTitle = page
 			.getByTestId('headerTitle')
 			.filter({hasText: 'Specification Facet Page'});
@@ -50,6 +55,9 @@ export class SpecificationFacetsPage {
 			.frameLocator('#modalIframe')
 			.getByTestId('searchOptionsFooter')
 			.getByRole('button', {exact: true, name: 'Save'});
+		this.selectSpecificationFacetPageInput = page
+			.getByTestId('selectLayout')
+			.getByLabel('Select Specification Facet Page');
 	}
 
 	async addSearchOptionsWidget() {
@@ -79,6 +87,26 @@ export class SpecificationFacetsPage {
 		await this.searchOptionsConfigurationSaveButton.click();
 	}
 
+	async deleteSpecificationPage() {
+		await this.selectSpecificationFacetPageInput.click();
+		await this.layoutsPage.deletePageButton.click();
+		await this.layoutsPage.deleteLayoutModal.waitFor({
+			state: 'attached',
+		});
+		await Promise.all([
+			this.layoutsPage.deleteLayoutModal.click(),
+			this.page.waitForResponse(
+				(resp) =>
+					resp.status() === 200 &&
+					resp
+						.url()
+						.includes(
+							'p_p_id=com_liferay_layout_admin_web_portlet_GroupPagesPortlet'
+						)
+			),
+		]);
+	}
+
 	async goto() {
 		await this.layoutsPage.goto();
 	}
@@ -86,7 +114,7 @@ export class SpecificationFacetsPage {
 	async goToPage() {
 		await this.layoutsPage.goToPages();
 		await Promise.all([
-			this.layoutsPage.pageLabel.click(),
+			this.pageLabel.click(),
 			this.page.waitForResponse(
 				(resp) =>
 					resp.status() === 200 &&
